requirements.txt
black==22.1.0
Django==4.0.3
django-environ==0.8.1
Pillow==9.0.1
cloudinary==1.29.0
django-cloudinary-storage==0.3.0
mutagen==1.45.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.20.0
